Add barrelchest anchor special attack
