+++ data/definitions/animation-definition-extras.yml
@@ -1305,3 +1305,5 @@ liquify:
   id: 6118
 backstab:
   id: 4198
+sunder:
+  id: 5870
