+++ data/definitions/graphic-definition-extras.yml
@@ -802,3 +802,5 @@ liquify:
   id: 1048
 backstab:
   id: 704
+sunder:
+  id: 1027
