+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/combat/melee/special/BarrelchestAnchor.kts
@@ -1,0 +1,84 @@
+package world.gregs.voidps.world.interact.entity.player.combat.melee.special
+
+import world.gregs.voidps.engine.entity.Registered
+import world.gregs.voidps.engine.entity.character.Character
+import world.gregs.voidps.engine.entity.character.contain.ItemChanged
+import world.gregs.voidps.engine.entity.character.player.Player
+import world.gregs.voidps.engine.entity.character.player.skill.Skill
+import world.gregs.voidps.engine.entity.character.update.visual.setAnimation
+import world.gregs.voidps.engine.entity.character.update.visual.setGraphic
+import world.gregs.voidps.engine.entity.item.EquipSlot
+import world.gregs.voidps.engine.entity.item.Item
+import world.gregs.voidps.engine.entity.item.equipped
+import world.gregs.voidps.engine.entity.set
+import world.gregs.voidps.engine.event.Priority
+import world.gregs.voidps.engine.event.on
+import world.gregs.voidps.world.interact.entity.combat.CombatHit
+import world.gregs.voidps.world.interact.entity.combat.CombatSwing
+import world.gregs.voidps.world.interact.entity.combat.hit
+import world.gregs.voidps.world.interact.entity.combat.weapon
+import world.gregs.voidps.world.interact.entity.player.combat.melee.specialAccuracyMultiplier
+import world.gregs.voidps.world.interact.entity.player.combat.melee.specialDamageMultiplier
+import world.gregs.voidps.world.interact.entity.player.combat.range.special.MAX_SPECIAL_ATTACK
+import world.gregs.voidps.world.interact.entity.player.combat.range.special.drainSpecialEnergy
+import world.gregs.voidps.world.interact.entity.player.combat.range.special.specialAttack
+
+fun isAnchor(item: Item?) = item != null && item.name == "barrelchest_anchor"
+
+on<Registered>({ isAnchor(it.equipped(EquipSlot.Weapon)) }) { player: Player ->
+    updateWeapon(player, player.equipped(EquipSlot.Weapon))
+}
+
+on<ItemChanged>({ container == "worn_equipment" && index == EquipSlot.Weapon.index && isAnchor(item) }) { player: Player ->
+    updateWeapon(player, item)
+}
+
+fun updateWeapon(player: Player, weapon: Item) {
+    player["attack_range"] = 1
+    player.weapon = weapon
+}
+
+on<CombatSwing>({ !swung() && isAnchor(it.weapon) }, Priority.LOW) { player: Player ->
+    player.setAnimation("anchor_attack")
+    player.hit(target)
+    delay = 6
+}
+
+on<CombatHit>({ isAnchor(weapon) }) { player: Player ->
+    player.setAnimation("anchor_block")
+}
+
+// Special attack
+
+specialAccuracyMultiplier(2.0, ::isAnchor)
+specialDamageMultiplier(1.1, ::isAnchor)
+
+on<CombatSwing>({ !swung() && it.specialAttack && isAnchor(it.weapon) }) { player: Player ->
+    if (!drainSpecialEnergy(player, MAX_SPECIAL_ATTACK / 2)) {
+        delay = -1
+        return@on
+    }
+    player.setAnimation("sunder")
+    player.setGraphic("sunder")
+    player.hit(target)
+    delay = 6
+}
+
+val skills = listOf(Skill.Defence, Skill.Strength, Skill.Prayer, Skill.Attack, Skill.Magic, Skill.Range)
+
+on<CombatHit>({ isAnchor(weapon) && special }) { character: Character ->
+    var drain = damage / 10
+    if (drain > 0) {
+        for (skill in skills) {
+            val current = character.levels.get(skill)
+            if (current <= 1) {
+                continue
+            }
+            character.levels.drain(skill, drain)
+            drain -= current
+            if (drain <= 0) {
+                break
+            }
+        }
+    }
+}
